Align clip-path checking more closely with spec
This change brings `clip-path` checking a bit closer into alignment with the
requirements in https://drafts.fxtf.org/css-masking-1/#the-clip-path.

Even with this change, the checker still doesn’t recognize circle(), ellipse()
or polygon() values for clip-path, but it does now recognize inset() values —
with the limitation of not yet recognizing the `round <‘border-radius’>`
optional argument.