Emit specific error for “star property hack” cases
This change augments the previous commit by making the parser emit a specific
error message for the case where an identifier is prefixed with an asterisk.

Having a specific error message for this case is more helpful to users than the
generic parse-error message the parser otherwise emits. (It also makes it
possible to filter the error messages in particular contexts — e.g., when
checking `style`-element contents with the HTML checker, it allows users to
catch those specific cases using the HTML-checker’s message-filtering features.)